refactoring code. moving _get_required_root_dataset, _get_subgroup, to common.py file
